commit rename folder to another folder

+++ BatchRename/MainWindow.xaml.cs
@@ -189,6 +189,28 @@ namespace BatchRename
             }
         }
 
+        private void MoveAllFilesToCopyFolder(string srcPath, string desPath)
+        {
+            if(Directory.Exists(srcPath) && Directory.Exists(desPath))
+            {
+                foreach (string dirPath in Directory.GetDirectories(srcPath, "*", SearchOption.AllDirectories))
+                {
+                    Directory.CreateDirectory(dirPath.Replace(srcPath, desPath));
+                }
+
+                foreach (string filePath in Directory.GetFiles(srcPath, "*.*", SearchOption.AllDirectories))
+                {
+                    System.IO.File.Copy(filePath, filePath.Replace(srcPath, desPath), true);
+                }
+            }
+        }
+
+        private bool IsEmptyFolder(string folderPath)
+        {
+            bool result = !Directory.EnumerateFileSystemEntries(folderPath).Any();
+            return result;
+        }
+
         private void btnStartBatchCopy_Click(object sender, RoutedEventArgs e)
         {
             int type = tcTargets.SelectedIndex;
@@ -205,7 +227,11 @@ namespace BatchRename
                 if (folderBrowserDialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
                 {
                     string directory = folderBrowserDialog.SelectedPath;
-
+                    if (!IsEmptyFolder(directory))
+                    {
+                        MessageBox.Show("Destination must be an empty folder");
+                        return;
+                    }
                     if (type == (int)FileType.File)
                     {
                         Dictionary<string, int> duplicationsCopyfile = new Dictionary<string, int>();
@@ -214,14 +240,14 @@ namespace BatchRename
                         {
                             string newPath = Path.Combine(directory, file.Name);
 
-                            try
+                            if (!System.IO.File.Exists(newPath))
                             {
                                 System.IO.File.Copy(
                                     file.Path,
                                     newPath
                                 );
                             }
-                            catch (Exception)
+                            else
                             {
                                 if (duplicationsCopyfile.ContainsKey(newPath))
                                 {
@@ -235,7 +261,7 @@ namespace BatchRename
                                 string copyNameWithDuplicate = $"{Path.GetFileNameWithoutExtension(file.Name)} ({duplicationsCopyfile[newPath]}) {Path.GetExtension(file.Name)}";
 
                                 newPath = Path.Combine(directory, copyNameWithDuplicate);
-                                
+
                                 System.IO.File.Copy(
                                     file.Path,
                                     newPath
@@ -281,42 +307,56 @@ namespace BatchRename
                     }
                     else
                     {
-                        Dictionary<string, int> duplications = new Dictionary<string, int>();
+                        Dictionary<string, int> duplicationsCopyFolder = new Dictionary<string, int>();
 
                         foreach (var folder in _folders)
                         {
                             string newPath = Path.Combine(directory, folder.Name);
 
-                            try
+                            if (!Directory.Exists(newPath))
                             {
                                 Directory.CreateDirectory(newPath);
+                                MoveAllFilesToCopyFolder(folder.Path, newPath);
                             }
-                            catch (Exception)
+                            else
                             {
-                                //duplcation folder
-                            }
-
-                            string newIdealName = Path.Combine(directory, folder.NewName);
-
-                            try
-                            {                                
-                                Directory.Move(folder.Path, newIdealName);
-                            }
-                            catch (Exception)
-                            {
-                                if (duplications.ContainsKey(newIdealName))
+                                if (duplicationsCopyFolder.ContainsKey(newPath))
                                 {
-                                    duplications[newIdealName]++;
+                                    duplicationsCopyFolder[newPath]++;
                                 }
                                 else
                                 {
-                                    duplications[newIdealName] = 1;
+                                    duplicationsCopyFolder[newPath] = 1;
+                                }
+                                string copyNameWithDuplicate = $"{Path.GetFileNameWithoutExtension(folder.Name)} ({duplicationsCopyFolder[newPath]})";
+                                newPath = Path.Combine(directory, copyNameWithDuplicate);
+
+                                Directory.CreateDirectory(newPath);
+                                MoveAllFilesToCopyFolder(folder.Path, newPath);
+                            }
+
+                            Dictionary<string, int> duplicationsNewFolderName = new Dictionary<string, int>();
+                            string newIdealName = Path.Combine(directory, folder.NewName);
+
+                            try
+                            {
+                                Directory.Move(newPath, newIdealName);
+                            }
+                            catch (Exception)
+                            {
+                                if (duplicationsNewFolderName.ContainsKey(newIdealName))
+                                {
+                                    duplicationsNewFolderName[newIdealName]++;
+                                }
+                                else
+                                {
+                                    duplicationsNewFolderName[newIdealName] = 1;
                                 }
 
-                                string newLessCollisionName = $"{Path.GetFileNameWithoutExtension(folder.NewName)} ({duplications[newIdealName]})";
+                                string newLessCollisionName = $"{Path.GetFileNameWithoutExtension(folder.NewName)} ({duplicationsNewFolderName[newIdealName]})";
 
                                 Directory.Move(
-                                    folder.Path,
+                                    newPath,
                                     Path.Combine(directory, newLessCollisionName)
                                 );
                             }
